django_project/classify_api/requirements.txt
appnope==0.1.0
backcall==0.1.0
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
Django==2.2.7
djangorestframework==3.10.3
imbalanced-learn==0.5.0
imblearn==0.0
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
joblib==0.14.0
kiwisolver==1.1.0
matplotlib==3.1.1
mdlp-discretization==0.3.2
numpy==1.17.4
pandas==0.25.3
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.5
pysbrl==0.4.2rc0
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.21.3
scipy==1.3.2
six==1.13.0
skope-rules==1.0.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.7
boto3==1.9.33
botocore==1.12.33